Refactoring of data components
