+++ EasyReflectometryApp/Gui/Components/ExperimentSimulationSetup.qml
@@ -19,7 +19,7 @@ Row {
     }
     EaElements.Parameter {
         id: xMin
-        enabled: !ExGlobals.Constants.proxy.experimentLoaded
+        enabled: !ExGlobals.Constants.proxy.data.experimentLoaded
         width: textFieldWidth()
         units: "Å<sup>-1</sup>"
         text: EaLogic.Utils.toFixed(ExGlobals.Constants.proxy.simulation.qRangeAsObj.x_min, 3)
@@ -34,7 +34,7 @@ Row {
     }
     EaElements.Parameter {
         id: xMax
-        enabled: !ExGlobals.Constants.proxy.experimentLoaded
+        enabled: !ExGlobals.Constants.proxy.data.experimentLoaded
         width: textFieldWidth()
         units: "Å<sup>-1</sup>"
         text: EaLogic.Utils.toFixed(ExGlobals.Constants.proxy.simulation.qRangeAsObj.x_max, 3)
@@ -49,7 +49,7 @@ Row {
     }
     EaElements.Parameter {
         id: xStep
-        enabled: !ExGlobals.Constants.proxy.experimentLoaded
+        enabled: !ExGlobals.Constants.proxy.data.experimentLoaded
         width: textFieldWidth()
         units: "Å<sup>-1</sup>"
         text: EaLogic.Utils.toFixed(ExGlobals.Constants.proxy.simulation.qRangeAsObj.x_step, 3)
